add likes

+++ src/main/java/com/example/dog_datting/repo/PostRepo.kt
@@ -3,6 +3,6 @@ package com.example.dog_datting.repo
 import com.example.dog_datting.db.Post
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface PostRepo : JpaRepository<Post, Int> {
+interface PostRepo : JpaRepository<Post, Long> {
     fun findByIdGreaterThanOrderByIdDesc(id: Long): List<Post>?
 }
